page cleanup
