+++ frontend/app/components/app/footer/AppFooterTitle.vue
