+++ frontend/app/components/common/AppCard.vue
