+++ frontend/app/pages/groups/index.vue
@@ -7,34 +7,26 @@
       <h1>Groups</h1>
       <div class="actions">
         <LoadingIndicator v-if="isLoading" />
-        <Button
-          label="Create group"
-          icon="pi pi-plus"
-          @click="showCreateGroup = true"
-        />
+        <CreateGroup @created="onCreated" />
       </div>
     </div>
     <GroupsList :query="query" />
-    <CreateGroup
-      v-if="showCreateGroup"
-      @close="closeCreateGroup"
-    />
   </AppCard>
 </template>
 
 <script setup lang="ts">
 import GroupsList from '~/components/groups/overview/list/GroupsList.vue';
-import CreateGroup from '~/components/groups/edit/CreateGroup.vue';
 import LoadingIndicator from '~/components/common/LoadingIndicator.vue';
-import AppCard from '~/components/app/AppCard.vue';
+import AppCard from '~/components/common/AppCard.vue';
+import CreateGroup from '~/components/groups/edit/CreateGroup.vue';
+import type { Query } from '~/types';
+import type { GetGroupsResponse } from '~/components/groups/overview/types';
 
-const query = await useLazyBackendApi('/Groups', { key: 'groups' });
-const isLoading = computed(() => query.status.value === 'pending');
+const query: Query<GetGroupsResponse> = await useLazyBackendApi('/Groups', { key: 'groups' });
+const isLoading = computed<boolean>(() => query.status.value === 'pending');
 
-const showCreateGroup = ref(false);
-const closeCreateGroup = (created: boolean) => {
-  showCreateGroup.value = false;
-  if (created) query.refresh();
+const onCreated = (): void => {
+  query.refresh();
 };
 </script>
